# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  controller.RoomCreateRequest:
    properties:
      capacity:
        type: integer
      floor:
        type: string
      images:
        type: string
      name:
        type: string
      type_id:
        type: integer
    required:
    - capacity
    - name
    - type_id
    type: object
  controller.RoomUpdateRequest:
    properties:
      capacity:
        type: integer
      floor:
        type: string
      images:
        type: string
      name:
        type: string
      status:
        type: integer
      type_id:
        type: integer
    type: object
...
  model.Room:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      description:
        type: string
      equipment:
        type: string
      floor:
        type: string
      id:
        type: integer
      images:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
      type:
        $ref: '#/definitions/model.RoomType'
      type_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.RoomType:
    properties:
      base_price:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      max_people:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应状态码
        type: integer
      data:
        description: 响应数据，omitempty表示为空时不序列化
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 棋室管理系统后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 棋室管理系统 API
  version: "1.0"
paths:
...
  /room:
    post:
      consumes:
      - application/json
      description: 管理员创建新房间
      parameters:
      - description: 房间信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RoomCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Room'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建房间
      tags:
      - 房间
  /room-types:
    get:
      consumes:
      - application/json
      description: 获取房间类型列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.RoomType'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取房间类型列表
      tags:
      - 房间类型
    post:
      consumes:
      - application/json
      description: 管理员创建房间类型
      parameters:
      - description: 房间类型信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RoomType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建房间类型
      tags:
      - 房间类型
  /room-types/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除房间类型
      parameters:
      - description: 房间类型ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除房间类型
      tags:
      - 房间类型
    get:
      consumes:
      - application/json
      description: 根据房间类型ID获取房间类型详情
      parameters:
      - description: 房间类型ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RoomType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取房间类型详情
      tags:
      - 房间类型
    put:
      consumes:
      - application/json
      description: 管理员更新房间类型信息
      parameters:
      - description: 房间类型ID
        in: path
        name: id
        required: true
        type: string
      - description: 房间类型更新信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RoomType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新房间类型
      tags:
      - 房间类型
  /rooms:
    get:
      consumes:
      - application/json
      description: 获取房间列表，支持按类型、楼层、状态筛选
      parameters:
      - description: 房间类型ID
        in: query
        name: type_id
        type: string
      - description: 楼层
        in: query
        name: floor
        type: string
      - description: 房间状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Room'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取房间列表
      tags:
      - 房间
  /rooms/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除房间
      parameters:
      - description: 房间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除房间
      tags:
      - 房间
    get:
      consumes:
      - application/json
      description: 根据房间ID获取房间详情
      parameters:
      - description: 房间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Room'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取房间详情
      tags:
      - 房间
    put:
      consumes:
      - application/json
      description: 管理员更新房间信息
      parameters:
      - description: 房间ID
        in: path
        name: id
        required: true
        type: string
      - description: 房间更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RoomUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Room'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新房间
      tags:
      - 房间
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"